+++ src/main/kotlin/com/github/srwi/pixellens/imageProviders/ImageProvider.kt
@@ -115,7 +115,7 @@ abstract class ImageProvider {
             "uint64" -> array.asType<Double>() / 256.0
             "int16", "int32" -> (array.asType<Float>() / 256f) + 128f
             "int64", "I" -> (array.asType<Double>() / 256.0) + 128.0
-            "float16", "float32" -> array.asType<Float>() * 255f
+            "float16", "bfloat16", "float32" -> array.asType<Float>() * 255f
             "float64", "F" -> array.asType<Double>() * 255.0
             "bool" -> array.asType<Float>() * 255f
             else -> throw IllegalArgumentException("Unsupported data type: $dataType")
@@ -141,6 +141,16 @@ abstract class ImageProvider {
                 mk.zeros<Float>(size).apply {
                     for (i in array.indices) {
                         this[i] = Float16.fromBits(array[i])
+                    }
+                }
+            }
+            "bfloat16" -> {
+                val size = bytes.size / 2
+                val array = ShortArray(size)
+                imageBuffer.asShortBuffer().get(array)
+                mk.zeros<Float>(size).apply {
+                    for (i in array.indices) {
+                        this[i] = BFloat16.fromBits(array[i])
                     }
                 }
             }
@@ -268,11 +278,21 @@ abstract class ImageProvider {
         }
     }
 
+    object BFloat16 {
+        fun fromBits(someBits: Short): Float {
+            val bits = (someBits.toInt() and 0xFFFF) shl 16
+            return Float.fromBits(bits)
+        }
+    }
+
     open fun shapeSupported(value: PyDebugValue): Boolean {
         val expression = value.evaluationExpression
         val shape = Python.evaluateExpression(value.frameAccessor, "$expression.shape").value ?: return false
         val shapeList = convertStringToShapeList(shape)
         if (shapeList.isEmpty()) return false
+
+        if (shapeList.any { it <= 0 }) return false
+        
         if (shapeList.size <= 4) return true
         for (i in 0 until (shapeList.size - 4)) {
             // Only dimensions of size 1 can be squeezed